Fix bug with default export generated when modules reference "this"
When the main module references "this", ESBuild generates a chunk
that has a default export at the end. When we then try to assemble
that into the final deduped bundle, we end up with an `export default`
in the middle of the wrapping IIFE, which is a syntax error.

This change looks for default exports in the main and worker modules
and, if it finds them, just changes them to a simple assignment to a
dummy variable.